Move message properties to ui-utils package
