+++ packages/rocketchat-ui-utils/tests/server.mocks.js
@@ -1,0 +1,9 @@
+import mock from 'mock-require';
+
+mock('meteor/rocketchat:emoji', {
+	emoji: {
+		list: {
+			
+		}
+	},
+});
